MDS-5268 Refactored document endpoints in docman
