+++ services/document-manager/backend/app/services/object_store_storage_service.py
@@ -43,13 +43,18 @@ class ObjectStoreStorageService():
             for chunk in iter(lambda: result['Body'].read(1048576), b''):
                 yield chunk
 
-        s3_response = self._client.get_object(Bucket=Config.OBJECT_STORE_BUCKET, Key=path)
+        print('hiiii')
+        print(Config.OBJECT_STORE_BUCKET, path, display_name)
+
+        s3_response = self._client.get_object(
+            Bucket=Config.OBJECT_STORE_BUCKET, Key=path)
         resp = Response(
             generate(s3_response),
             mimetype='application/pdf' if '.pdf' in display_name.lower() else 'application/zip',
             headers={
                 'Content-Disposition':
-                ('attachment; ' if as_attachment else '') + ('filename=' + display_name)
+                ('attachment; ' if as_attachment else '') +
+                ('filename=' + display_name)
             })
         return resp
 
@@ -80,7 +85,8 @@ class ObjectStoreStorageService():
         s3_etag = self.s3_etag(key)
         fs_etag = fs_etag if fs_etag else self.calculate_s3_etag(filename)
         if (s3_etag != fs_etag):
-            raise Exception('ETag of the uploaded file and local file do not match!')
+            raise Exception(
+                'ETag of the uploaded file and local file do not match!')
 
         return True, key
 
@@ -106,14 +112,16 @@ class ObjectStoreStorageService():
 
     def copy_file(self, source_key, key):
         copy_source = {'Bucket': Config.OBJECT_STORE_BUCKET, 'Key': source_key}
-        self._client.copy(CopySource=copy_source, Bucket=Config.OBJECT_STORE_BUCKET, Key=key)
+        self._client.copy(CopySource=copy_source,
+                          Bucket=Config.OBJECT_STORE_BUCKET, Key=key)
 
     def delete_file(self, key):
         self._client.delete_object(Bucket=Config.OBJECT_STORE_BUCKET, Key=key)
 
     def file_exists(self, key):
         try:
-            self._client.head_object(Bucket=Config.OBJECT_STORE_BUCKET, Key=key)
+            self._client.head_object(
+                Bucket=Config.OBJECT_STORE_BUCKET, Key=key)
             return True
         except ClientError as e:
             if (int(e.response['Error']['Code']) == 404):
